Load Information
An update has been made to the tours page. Now the page loads in a list
of buttons with the name of each tour. When a user clicks a tour, a new
page is loaded showing a description of the tour. Map interface still
to be loaded into here and linking tour to AR still to be complete.
The preferences page now loads in a list of points of interest. These
are just text for now which doesn't do anything.