Changes to Preferences and Tours
The change to the preferences page is that it now displays a list of
poi's as toggle boxes. Each toggle box can be selected and deselected.
When the 'show on map' button is pressed, the console will display the
information about that poi. In the future, this will be passed to the
AR camera for each POI.
The tours page hasn't changed appearance wise, but now when a tour is
selected, the poi's that would be on that tour are now listed in the
console. Again, once the AR side is being completed, information can
now be passed onto the AR camera.
Another change made was just a tidy up to the code. Moving the split
strings method to a class of its own as it is used across many classes
and also taking the setters out of the GetPois and GetTours classes as
they weren't needed.