New AR Scenes
Two new AR scenes created for preferences and tours. These new scenes
have been created without the AR being implemented just now. The
information that is to be passed to them by the previous scene (tour
information for tours and the users selected preferences for
preferences) has been implemented so now the correct information can
later be displayed on the AR scene.